refactored 2/3 menus to make our code nicer
