+++ src/proj7ZhouRinkerSahChistolini/Controllers/Controller.java
@@ -12,18 +12,8 @@
 package proj7ZhouRinkerSahChistolini.Controllers;
 
 import javafx.fxml.FXML;
-import javafx.application.Platform;
 import javafx.scene.control.MenuItem;
-import proj7ZhouRinkerSahChistolini.Controllers.Actions.DeleteNoteAction;
-import proj7ZhouRinkerSahChistolini.Controllers.Actions.PasteAction;
-import proj7ZhouRinkerSahChistolini.Controllers.Actions.SelectAction;
 import proj7ZhouRinkerSahChistolini.Models.Instrument;
-import proj7ZhouRinkerSahChistolini.Views.SelectableRectangle;
-
-
-import java.awt.datatransfer.UnsupportedFlavorException;
-import java.io.IOException;
-import java.util.Collection;
 
 /**
  * Handles menu GUI interactions with other controllers
@@ -38,11 +28,14 @@ public class Controller {
     @FXML
     private InstrumentPanelController instrumentPaneController;
 
+    /** an editMenuController field which allows us to talk to the editMenu */
+    @FXML
+    private EditMenuController editMenuController;
+
     /** a controller to assist in bindings between the menus and controllers*/
     private BindingController bindingController;
 
     private ClipBoardController clipboardController;
-
 
     /**
      * All of our MenuItem are put into fields
@@ -54,24 +47,6 @@ public class Controller {
     private MenuItem stopButton;
     @FXML
     public MenuItem startButton;
-    @FXML
-    public MenuItem selectAllButton;
-    @FXML
-    private MenuItem deleteButton;
-    @FXML
-    private MenuItem groupButton;
-    @FXML
-    private MenuItem ungroupButton;
-    @FXML
-    private MenuItem undoButton;
-    @FXML
-    private MenuItem redoButton;
-    @FXML
-    private MenuItem cutButton;
-    @FXML
-    private MenuItem copyButton;
-    @FXML
-    private MenuItem pasteButton;
 
     /** Initializes the controllers so they can communicate properly */
     @FXML
@@ -84,6 +59,9 @@ public class Controller {
                 this,
                 this.compositionPanelController,
                 this.clipboardController);
+        this.editMenuController.init(this.compositionPanelController,
+                                     this.bindingController,
+                                     this.clipboardController);
         bindMenuItems();
     }
 
@@ -92,113 +70,14 @@ public class Controller {
         return this.instrumentPaneController.getSelectedInstrument();
     }
 
-    /**
-     * Ensures that all processes are killed on the
-     * destruction of the window.
-     */
-    @FXML
-    public void cleanUpOnExit() {
-        Platform.exit();
-        System.exit(0);
-    }
-
     @FXML
     public void playComposition() { this.compositionPanelController.playComposition(); }
 
     @FXML
     public void stopComposition() { this.compositionPanelController.stopComposition(); }
 
-    @FXML
     /**
-     * deletes the selected notes
-     */
-    public void deleteSelectedNotes() {
-        DeleteNoteAction deletedNotes = new DeleteNoteAction(
-            this.compositionPanelController.getSelectedRectangles(),
-            this.compositionPanelController.getSelectedNotes(),
-            this.compositionPanelController
-        );
-        this.compositionPanelController.deleteSelectedNotes();
-        this.compositionPanelController.addAction(deletedNotes);
-    }
-
-    @FXML
-    /**
-     * selects all the notes
-     */
-    public void selectAllNotes() {
-        //add
-        Collection<SelectableRectangle> before = (
-            this.compositionPanelController.getSelectedRectangles()
-        );
-        this.compositionPanelController.selectAllNotes();
-        this.compositionPanelController.addAction(
-                new SelectAction(
-                        before,
-                        this.compositionPanelController.getRectangles(),
-                        this.compositionPanelController
-                )
-        );
-    }
-
-    @FXML
-    /**
-     * group the selected notes
-     */
-    public void groupSelected() { this.compositionPanelController.groupSelected(this.compositionPanelController.getSelectedRectangles()); }
-
-    @FXML
-    /**
-     * copies the selected rectangles
-     */
-    public void copySelected() {
-        this.clipboardController.copySelected();
-    }
-
-    @FXML
-    /**
-     * cut the selected rectangles
-     */
-    public void cutSelected() {
-        this.clipboardController.copySelected();
-        this.deleteSelectedNotes();
-    }
-
-    @FXML
-    /**
-     * paste the copied rectangles
-     */
-    public void pasteSelected() throws IOException, UnsupportedFlavorException {
-        this.clipboardController.pasteSelected();
-        PasteAction pastedNotes = new PasteAction(this.compositionPanelController.getSelectedRectangles(),
-                this.compositionPanelController.getSelectedNotes(), this.compositionPanelController);
-        this.compositionPanelController.addAction(pastedNotes);
-    }
-
-    @FXML
-    /**
-     * ungroup the selected notes
-     */
-    public void ungroupSelected() { this.compositionPanelController.ungroupSelected(this.compositionPanelController.getSelectedRectangles()); }
-
-    @FXML
-    /**
-     * undoes the latest action
-     */
-    public void undo() {
-        this.compositionPanelController.getActionController().undo();
-    }
-
-    @FXML
-    /**
-     * redoes the latest undo action
-     */
-    public void redo() {
-        this.compositionPanelController.getActionController().redo();
-    }
-
-    /**
-     * returns whether or not the composition is playing
+     * Sets up the bindings for the menuButtons in order to disable them
      */
     public void bindMenuItems() {
         //stopButton
@@ -208,42 +87,6 @@ public class Controller {
         //startButton
         this.startButton.disableProperty().bind(
                 this.bindingController.getChildrenProperty().sizeProperty().isEqualTo(0)
-        );
-        //selectAllButton
-        this.selectAllButton.disableProperty().bind(
-                this.bindingController.getChildrenProperty().sizeProperty().isEqualTo(0)
-        );
-        //deleteButton
-        this.deleteButton.disableProperty().bind(
-                this.bindingController.getAreNotesSelectedBinding()
-        );
-        //groupButton
-        this.groupButton.disableProperty().bind(
-                this.bindingController.getMultipleSelectedBinding()
-        );
-        //UngroupButton
-        this.ungroupButton.disableProperty().bind(
-                this.bindingController.getGroupSelectedBinding()
-        );
-        //redoButton
-        this.redoButton.disableProperty().bind(
-                this.bindingController.getRedoEmptyBinding()
-        );
-        //undoButton
-        this.undoButton.disableProperty().bind(
-                this.bindingController.getUndoEmptyBinding()
-        );
-        //cutButton
-        this.cutButton.disableProperty().bind(
-                this.bindingController.getAreNotesSelectedBinding()
-        );
-        //copyButton
-        this.copyButton.disableProperty().bind(
-                this.bindingController.getAreNotesSelectedBinding()
-        );
-        //pasteButton
-        this.pasteButton.disableProperty().bind(
-                this.bindingController.getClipBoardEmptyBinding()
         );
     }
 }
