+++ src/proj7ZhouRinkerSahChistolini/Views/editMenu.fxml
@@ -1,0 +1,56 @@
+<!-- /**
+ * File: editMenu.fxml
+ * @author Victoria Chistolini
+ * @author Edward Zhou
+ * @author Alex Rinker
+ * @author Vivek Sah
+ * Class: CS361
+ * Project: 8
+ * Date: November 14, 2016
+ */ -->
+
+<?import javafx.scene.control.*?>
+
+<Menu text="Edit"
+      xmlns:fx="http://javafx.com/fxml"
+      fx:controller="proj7ZhouRinkerSahChistolini.Controllers.EditMenuController">
+    <MenuItem fx:id="undoButton"
+              text="Undo"
+              onAction="#undo"
+              accelerator="Shortcut+Z"/>
+    <MenuItem fx:id="redoButton"
+              text="Redo"
+              onAction="#redo"
+              accelerator="Shortcut+SHIFT+Z"/>
+    <SeparatorMenuItem/>
+    <MenuItem fx:id="selectAllButton"
+              text="Select All"
+              onAction="#selectAllNotes"
+              accelerator="Shortcut+A"/>
+    <MenuItem fx:id="deleteButton"
+              text="Delete"
+              onAction="#deleteSelectedNotes"
+              accelerator="Shortcut+BACKSPACE"/>
+    <SeparatorMenuItem/>
+    <MenuItem fx:id="groupButton"
+              text="Group"
+              onAction="#groupSelected"
+              accelerator="Shortcut+G"/>
+    <MenuItem fx:id="ungroupButton"
+              text="Ungroup"
+              onAction="#ungroupSelected"
+              accelerator="Shortcut+U"/>
+    <SeparatorMenuItem/>
+    <MenuItem fx:id="cutButton"
+              text="Cut"
+              onAction="#cutSelected"
+              accelerator="Shortcut+X"/>
+    <MenuItem fx:id="copyButton"
+              text="Copy"
+              onAction="#copySelected"
+              accelerator="Shortcut+C"/>
+    <MenuItem fx:id="pasteButton"
+              text="Paste"
+              onAction="#pasteSelected"
+              accelerator="Shortcut+V"/>
+</Menu>
