+++ src/proj7ZhouRinkerSahChistolini/Views/fileMenu.fxml
@@ -1,0 +1,18 @@
+<!-- /**
+ * File: fileMenu.fxml
+ * @author Victoria Chistolini
+ * @author Edward Zhou
+ * @author Alex Rinker
+ * @author Vivek Sah
+ * Class: CS361
+ * Project: 8
+ * Date: November 14, 2016
+ */ -->
+
+<?import javafx.scene.control.*?>
+
+<Menu text="File"
+      xmlns:fx="http://javafx.com/fxml"
+      fx:controller="proj7ZhouRinkerSahChistolini.Controllers.FileMenuController">
+    <MenuItem text="Exit" onAction="#cleanUpOnExit" accelerator="Shortcut+Q"/>
+</Menu>
