+++ src/proj7ZhouRinkerSahChistolini/Views/main.fxml
@@ -17,50 +17,9 @@
             fx:controller="proj7ZhouRinkerSahChistolini.Controllers.Controller">
     <top>
         <MenuBar>
-            <Menu text="File">
-                <MenuItem text="Exit" onAction="#cleanUpOnExit" accelerator="Shortcut+Q"/>
-            </Menu>
-            <Menu text="Edit">
-                <MenuItem fx:id="undoButton"
-                          text="Undo"
-                          onAction="#undo"
-                          accelerator="Shortcut+Z"/>
-                <MenuItem fx:id="redoButton"
-                          text="Redo"
-                          onAction="#redo"
-                          accelerator="Shortcut+SHIFT+Z"/>
-                <SeparatorMenuItem/>
-                <MenuItem fx:id="selectAllButton"
-                          text="Select All"
-                          onAction="#selectAllNotes"
-                          accelerator="Shortcut+A"/>
-                <MenuItem fx:id="deleteButton"
-                          text="Delete"
-                          onAction="#deleteSelectedNotes"
-                          accelerator="Shortcut+BACKSPACE"/>
-                <SeparatorMenuItem/>
-                <MenuItem fx:id="groupButton"
-                          text="Group"
-                          onAction="#groupSelected"
-                          accelerator="Shortcut+G"/>
-                <MenuItem fx:id="ungroupButton"
-                          text="Ungroup"
-                          onAction="#ungroupSelected"
-                          accelerator="Shortcut+U"/>
-                <SeparatorMenuItem/>
-                <MenuItem fx:id="cutButton"
-                          text="Cut"
-                          onAction="#cutSelected"
-                          accelerator="Shortcut+X"/>
-                <MenuItem fx:id="copyButton"
-                          text="Copy"
-                          onAction="#copySelected"
-                          accelerator="Shortcut+C"/>
-                <MenuItem fx:id="pasteButton"
-                          text="Paste"
-                          onAction="#pasteSelected"
-                          accelerator="Shortcut+V"/>
-            </Menu>
+            <fx:include source="fileMenu.fxml"/>
+            <fx:include fx:id="editMenu"
+                        source="editMenu.fxml"/>
             <Menu text="Action">
                 <MenuItem fx:id="startButton"
                           text="Play"
